Combine PromisedPlayedExoPlayer with handler
